Feat(#3): Added caching of all weapons, their stats and all their skins
Moved reward types from being hard coded strings fragmented all over the project to a centralized RewardType object

BREAKING CHANGE: Changed local cache database structure

+++ app/src/main/java/dev/bittim/valolink/main/data/local/game/entity/weapon/stats/WeaponAdsStatsEntity.kt
@@ -9,34 +9,29 @@ import dev.bittim.valolink.main.domain.model.game.weapon.stats.WeaponAdsStats
 
 @Entity(
     tableName = "WeaponAdsStats",
-    foreignKeys = [
-        ForeignKey(
-            entity = WeaponStatsEntity::class,
-            parentColumns = ["uuid"],
-            childColumns = ["weaponStats"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ],
-    indices = [
-        Index(
-            value = ["uuid"],
-            unique = true
-        ),
-        Index(
-            value = ["weaponStats"],
-            unique = true
-        )
-    ]
+    foreignKeys = [ForeignKey(
+        entity = WeaponStatsEntity::class,
+        parentColumns = ["uuid"],
+        childColumns = ["weaponStats"],
+        onDelete = ForeignKey.CASCADE
+    )],
+    indices = [Index(
+        value = ["uuid"],
+        unique = true
+    ), Index(
+        value = ["weaponStats"],
+        unique = true
+    )]
 )
 data class WeaponAdsStatsEntity(
     @PrimaryKey val uuid: String,
     override val version: String,
     val weaponStats: String,
-    val zoomMultiplier: Float,
-    val fireRate: Float,
-    val runSpeedMultiplier: Float,
+    val zoomMultiplier: Double,
+    val fireRate: Double,
+    val runSpeedMultiplier: Double,
     val burstCount: Int,
-    val firstBulletAccuracy: Float,
+    val firstBulletAccuracy: Double,
 ) : GameEntity() {
     fun toType(): WeaponAdsStats {
         return WeaponAdsStats(
