+++ app/src/main/java/dev/bittim/valolink/main/data/local/game/entity/weapon/stats/WeaponAirBurstStatsEntity.kt
@@ -9,31 +9,26 @@ import dev.bittim.valolink.main.domain.model.game.weapon.stats.WeaponAirBurstSta
 
 @Entity(
     tableName = "WeaponAirBurstStats",
-    foreignKeys = [
-        ForeignKey(
-            entity = WeaponStatsEntity::class,
-            parentColumns = ["uuid"],
-            childColumns = ["weaponStats"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ],
-    indices = [
-        Index(
-            value = ["uuid"],
-            unique = true
-        ),
-        Index(
-            value = ["weaponStats"],
-            unique = true
-        )
-    ]
+    foreignKeys = [ForeignKey(
+        entity = WeaponStatsEntity::class,
+        parentColumns = ["uuid"],
+        childColumns = ["weaponStats"],
+        onDelete = ForeignKey.CASCADE
+    )],
+    indices = [Index(
+        value = ["uuid"],
+        unique = true
+    ), Index(
+        value = ["weaponStats"],
+        unique = true
+    )]
 )
 data class WeaponAirBurstStatsEntity(
     @PrimaryKey val uuid: String,
     override val version: String,
     val weaponStats: String,
     val shotgunPelletCount: Int,
-    val burstDistance: Float,
+    val burstDistance: Double,
 ) : GameEntity() {
     fun toType(): WeaponAirBurstStats {
         return WeaponAirBurstStats(
