+++ app/src/main/java/dev/bittim/valolink/main/data/local/game/entity/weapon/stats/WeaponDamageRangeEntity.kt
@@ -9,34 +9,29 @@ import dev.bittim.valolink.main.domain.model.game.weapon.stats.WeaponDamageRange
 
 @Entity(
     tableName = "WeaponDamageRanges",
-    foreignKeys = [
-        ForeignKey(
-            entity = WeaponStatsEntity::class,
-            parentColumns = ["uuid"],
-            childColumns = ["weaponStats"],
-            onDelete = ForeignKey.CASCADE
-        )
-    ],
-    indices = [
-        Index(
-            value = ["uuid"],
-            unique = true
-        ),
-        Index(
-            value = ["weaponStats"],
-            unique = true
-        )
-    ]
+    foreignKeys = [ForeignKey(
+        entity = WeaponStatsEntity::class,
+        parentColumns = ["uuid"],
+        childColumns = ["weaponStats"],
+        onDelete = ForeignKey.CASCADE
+    )],
+    indices = [Index(
+        value = ["uuid"],
+        unique = true
+    ), Index(
+        value = ["weaponStats"],
+        unique = true
+    )]
 )
 data class WeaponDamageRangeEntity(
     @PrimaryKey val uuid: String,
     override val version: String,
     val weaponStats: String,
-    val rangeStartMeters: Float,
-    val rangeEndMeters: Float,
-    val headDamage: Float,
-    val bodyDamage: Float,
-    val legDamage: Float,
+    val rangeStartMeters: Double,
+    val rangeEndMeters: Double,
+    val headDamage: Double,
+    val bodyDamage: Double,
+    val legDamage: Double,
 ) : GameEntity() {
     fun toType(): WeaponDamageRange {
         return WeaponDamageRange(
