+++ app/src/main/java/dev/bittim/valolink/main/domain/model/game/weapon/stats/WeaponAirBurstStats.kt
@@ -2,5 +2,5 @@ package dev.bittim.valolink.main.domain.model.game.weapon.stats
 
 data class WeaponAirBurstStats(
     val shotgunPelletCount: Int,
-    val burstDistance: Float,
+    val burstDistance: Double,
 )
