+++ app/src/main/java/dev/bittim/valolink/main/domain/model/game/weapon/stats/WeaponAltShotgunStats.kt
@@ -2,5 +2,5 @@ package dev.bittim.valolink.main.domain.model.game.weapon.stats
 
 data class WeaponAltShotgunStats(
     val shotgunPelletCount: Int,
-    val burstRate: Float,
+    val burstRate: Double,
 )
