+++ app/src/main/java/dev/bittim/valolink/main/domain/model/game/weapon/stats/WeaponDamageRange.kt
@@ -1,9 +1,9 @@
 package dev.bittim.valolink.main.domain.model.game.weapon.stats
 
 data class WeaponDamageRange(
-    val rangeStartMeters: Float,
-    val rangeEndMeters: Float,
-    val headDamage: Float,
-    val bodyDamage: Float,
-    val legDamage: Float,
+    val rangeStartMeters: Double,
+    val rangeEndMeters: Double,
+    val headDamage: Double,
+    val bodyDamage: Double,
+    val legDamage: Double,
 )
